EndpointManager: fix deadlock when releasing an endpoint
[ upstream commit 061e55f3fe7f9e0028c7fa779997c22ce9670ae7 ]

In high-churn clusters, there can be a three-party deadlock between the
EndpointManager, the PolicyRepository, and a given Endpoint. One of the
"links in the chain" is merely trying to get the container ID and
namespace+name of an Endpoint for logging. Which we already have.

So, rather than trying to lock an Endpoint to get it's identifiers
again, just use the copy we already have.

Fixes: dae07b58 (endpointmanager: Remove goroutine for ID release)
Signed-off-by: Casey Callendrello <cdc@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>