ctmap: refactor NAT GC logic
[ upstream commit 183726c8da72d899ef9c205bb7f80d1a2577c099 ]

After adding support to the GC for NAT entries created by the DSR logic,
we ended up handling twice the case for orphan egress entries as
currently when we find an orphan ingress one we derive the related
egress one and delete it as well.

Since this case is already handled by the logic that goes through
egress NAT entries, simply remove it.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>