gha: Add gateway API conformance test
This commit is add a simple gateway api conformance test from
upstream. The goal is to have tests running in pipeline, then
we will iterate steadily.

The upstream conformance, by default, is configured with alphav1
feature ReferenceGrant. To reduce the coupling with upstream, the
conformance_test.go is added for flexibility, for example, query
param tests are enabled.

Signed-off-by: Tam Mach <tam.mach@cilium.io>